[YS-462] refactor: 체크 상태에 따라 페이지 전체가 리렌더링되는 부분 컴포넌트 분리
