+++ src/app/join/mobile/Researcher/UnivEmailStep/UnivEmailStep.tsx
@@ -1,11 +1,11 @@
 import { FormProvider, useFormContext, useWatch } from 'react-hook-form';
 
+import NextButton from './NextButton';
 import ServiceAgreeBottomSheet from '../../components/ServiceAgreeBottomSheet/ServiceAgreeBottomSheet';
 import TitleSection from '../../components/TitleSection/TitleSection';
-import { bottomButtonLayout, mainContentLayout } from '../../page.css';
+import { mainContentLayout } from '../../page.css';
 import UnivEmailInputContainer from '../UnivEmailInputContainer/UnivEmailInputContainer';
 
-import Button from '@/components/Button/Button';
 import useOverlay from '@/hooks/useOverlay';
 import { ResearcherJoinSchemaType } from '@/schema/join/ResearcherJoinSchema';
 
@@ -14,18 +14,9 @@ interface UnivEmailStepProps {
 }
 
 const UnivEmailStep = ({ onNext }: UnivEmailStepProps) => {
-  const form = useFormContext<ResearcherJoinSchemaType>();
-  const { control } = form;
-
   const { open, close } = useOverlay();
-
-  // 버튼 활성화 조건
-  const isEmailVerified = useWatch({ name: 'isEmailVerified', control });
-  const isTermOfService = useWatch({ name: 'isTermOfService', control });
-  const isPrivacy = useWatch({ name: 'isPrivacy', control });
-
-  const isValidCheck = isTermOfService && isPrivacy;
-  const canNext = isEmailVerified && isValidCheck;
+  const form = useFormContext<ResearcherJoinSchemaType>();
+  const isEmailVerified = useWatch({ name: 'isEmailVerified', control: form.control });
 
   const openServiceAgreeBottomSheet = () => {
     open(() => (
@@ -50,17 +41,9 @@ const UnivEmailStep = ({ onNext }: UnivEmailStepProps) => {
       {/* 학교 메일 인증 */}
       <UnivEmailInputContainer openServiceAgreeBottomSheet={openServiceAgreeBottomSheet} />
 
+      {/* 다음 버튼 */}
       {isEmailVerified && (
-        <div className={bottomButtonLayout}>
-          <Button
-            variant="primary"
-            size="small"
-            height="56px"
-            onClick={canNext ? onNext : openServiceAgreeBottomSheet}
-          >
-            다음
-          </Button>
-        </div>
+        <NextButton onNext={onNext} openServiceAgreeBottomSheet={openServiceAgreeBottomSheet} />
       )}
     </main>
   );
